+++ lib/sidekiq/capistrano.rb
@@ -1,16 +1,19 @@
 Capistrano::Configuration.instance.load do
+  before "deploy", "sidekiq:quiet"
   after "deploy", "sidekiq:restart"
+
+  _cset(:sidekiq_timeout) { 5 }
 
   namespace :sidekiq do
 
-    desc "Force stop sidekiq"
-    task :kill do
-      run "cd #{current_path} && kill `cat tmp/pids/sidekiq.pid` && sleep 5 && kill -9 `cat tmp/pids/sidekiq.pid`"
+    desc "Quiet sidekiq (stop accepting new work)"
+    task :quiet do
+      run "cd #{current_path} && kill -USR1 `cat tmp/pids/sidekiq.pid`"
     end
 
     desc "Stop sidekiq"
     task :stop do
-      run "cd #{current_path} && kill `cat tmp/pids/sidekiq.pid`"
+      run "cd #{current_path} && kill `cat tmp/pids/sidekiq.pid` && sleep #{fetch :sidekiq_timeout} && kill -9 `cat tmp/pids/sidekiq.pid` && rm tmp/pids/sidekiq.pid"
     end
 
     desc "Start sidekiq"